requirements.txt
torch==2.8.0
torchaudio==2.8.0
numpy==2.4.4
scipy==1.16.1
soundfile==0.13.1
librosa==0.11.0
faster-whisper==1.2.1
transformers==4.51.3
sentencepiece==0.2.1
speechbrain==1.1.0
coqui-tts==0.26.2
deepfilternet==0.5.6
pymcd==0.2.1
pyworld==0.3.5
phonemizer==3.3.0
scikit-learn==1.7.1
jiwer==4.0.0
tqdm==4.67.1
